refactor(worker): simplify to mvp heartbeat-only design
- Replace register/heartbeat/unregister pattern with single upsert heartbeat
- Remove cleanup logic from CLI worker command
- Heartbeat API auto-creates/updates workers using deterministic IDs
- Workers naturally expire after 60s without heartbeat
- Reduces API surface from 3 endpoints to 1
